% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Accuracy
}
\description{
Compute Accuracy using confusion matrix.
Accuracy the describes the fraction of correctly classified instances in the test set.
}
