% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_crossentropy.R
\name{binary_crossentropy}
\alias{binary_crossentropy}
\title{Binary Crossentropy}
\usage{
binary_crossentropy(y_real, y_predicted)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicte values (probabiblities by class).}
}
\value{
integer value of Binary Crossentropy
}
\description{
Compute Binary Crossentropy.
The squared loss, in a sense, quantifies the error (or alternatively closeness) of the predicted label to the true label.
When specialized to the case of probabilistic classifiers, this then can be interpreted as a reliability measure.
This function assumes that inputs are given as probability vectors (and not matrices).
}
