% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_crossentropy.R
\name{categorical_crossentropy}
\alias{categorical_crossentropy}
\title{Categorical Crossentropy}
\usage{
categorical_crossentropy(y_real, y_predicted)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicte values (probabiblities by class).}
}
\value{
integer value of Categorical Crossentropy
}
\description{
Compute Categorical Crossentropy.
The squared loss, in a sense, quantifies the error (or alternatively closeness) of the predicted label to the true label.
When specialized to the case of probabilistic classifiers, this then can be interpreted as a reliability measure.
}
