% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ratio.R
\name{class_ratio}
\alias{class_ratio}
\title{Class Ratio}
\usage{
class_ratio(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Class Ratio
}
\description{
Compute Class Ratio using confusion matrix.
}
