% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_kappa.R
\name{cohen_kappa}
\alias{cohen_kappa}
\title{Cohen's Kappa}
\usage{
cohen_kappa(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer Cohen's Kappa value.
}
\description{
Calculate Cohen's Kappa using confusion matrix.
Contrary to Scott's Pi, here the chance agreement is defined considering the two processes to be fixed.
For more information see Cohen (1960).
}
