% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate.R
\name{error_rate}
\alias{error_rate}
\title{Epirical risk}
\usage{
error_rate(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Error Rate
}
\description{
Compute Error Rate using confusion matrix.
Error rate measures the fraction of the instances from the test set that are misclassified by the learning algorithm.
This measurement further includes the instances from all classes.
}
