% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_measure.R
\name{f_measure}
\alias{f_measure}
\title{F-measure}
\usage{
f_measure(confusion_matrix, alpha = 1)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}

\item{alpha}{Alpha value for weighted harmonic mean calculation.
Defaults to 1, which produces a balanced F-measure.}
}
\value{
integer value of F-measure
}
\description{
Compute F-measure using confusion matrix.
}
