% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_negative.R
\name{false_negative}
\alias{false_negative}
\title{False Negative}
\usage{
false_negative(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of False Negatives count
}
\description{
Count False Negatives using confusion matrix.
}
