% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_negative_rate.R
\name{false_negative_rate}
\alias{false_negative_rate}
\title{False Negative Rate}
\usage{
false_negative_rate(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of False Negatives Rate values.
}
\description{
Calculate False Negative Rate using confusion matrix.
}
