% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_positive.R
\name{false_positive}
\alias{false_positive}
\title{False Positive}
\usage{
false_positive(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of False Positives count
}
\description{
Count False Positives using confusion matrix.
}
