% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_positive_rate.R
\name{false_positive_rate}
\alias{false_positive_rate}
\title{False Positive Rate}
\usage{
false_positive_rate(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of False Positives Rate values.
}
\description{
Calculate False Positive Rate using confusion matrix.
Proportion of instances assigned to class i of interest that actually do not belong to this class.
}
