% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric Mean}
\usage{
geometric_mean(confusion_matrix, type = "tnr")
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}

\item{type}{Specifies whether True Negative Rate ("tnr") or Precision ("prec") should be used for calculation of the geometric mean.
Defaults to "tnr".}
}
\value{
integer value of Geometric Mean
}
\description{
Compute Geometric Mean using confusion matrix.
}
