% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ratio.R
\name{likelihood_ratio}
\alias{likelihood_ratio}
\title{Likelihood Ratio}
\usage{
likelihood_ratio(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
matrix of Likelihood Ratio values.
}
\description{
Likelihood Ratio using confusion matrix.
An important measure related to the sensitivity and specificity of the classifier.
A higher positive likelihood and a lower negative likelihood mean better performance on positive and negative classes, respectively (Japkowitcz, 2011).
For more information Deeks and Altman (2004).
}
