% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error (MAE)}
\usage{
mae(y_real, y_predicted, rowsums = TRUE)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicted values (probabiblities by class).}

\item{rowsums}{Whether in case of multiple classes in dataframe the metrics should be summed across classes before calculating the mean.}
}
\value{
integer value of Mean Absolute Error (MAE)
}
\description{
Compute Mean Absolute Error (MAE).
#'
}
