% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error (MAPE)}
\usage{
mape(y_real, y_predicted)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicte values (probabiblities by class).}
}
\value{
integer value of Mean Absolute Percentage Error (MAPE)
}
\description{
Compute Mean Absolute Percentage Error (MAPE).
This metric is not suitable for discrete analysis and probably will be removed from the package contents in near future.
}
