% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Square Error (MSE)}
\usage{
mse(y_real, y_predicted, rowsums = TRUE)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicte values (probabiblities by class).}

\item{rowsums}{Whether in case of multiple classes in dataframe the metrics should be summed across classes before calculating the mean.}
}
\value{
integer value of Mean Square Error (MSE)
}
\description{
Compute Mean Square Error (MSE).
The squared loss, in a sense, quantifies the error (or alternatively closeness) of the predicted label to the true label.
When specialized to the case of probabilistic classifiers, this then can be interpreted as a reliability measure.
}
