% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_predicted_values.R
\name{negative_predicted_values}
\alias{negative_predicted_values}
\title{Negative Predicted Values}
\usage{
negative_predicted_values(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Negative Predicted Values
}
\description{
Compute Negative Predicted Values using confusion matrix.
}
