\name{performancer-package}
\alias{performancer-package}
\alias{performancer}
\docType{package}
\title{
\packageTitle{performancer}
}
\description{
\packageDescription{performancer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{performancer}
\packageIndices{performancer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{performancer}

Maintainer: \packageMaintainer{performancer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
