% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_coordinates.R
\name{plot.pr_coordinates}
\alias{plot.pr_coordinates}
\title{Plot PR coordinates}
\usage{
\method{plot}{pr_coordinates}(pr_coordinates, ...)
}
\arguments{
\item{pr_coordinates}{A PR coordinates object.
A simple data.frame containing respective `x` and `y` coordinates.
The coordinates should be PR space compliant.}
}
\value{
plot of PR coordinates for given confusion matrix
}
\description{
Plot PR points using provided coordinates.
}
