% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve.R
\name{plot.roc_curve}
\alias{plot.roc_curve}
\title{Plot ROC coordinates}
\usage{
\method{plot}{roc_curve}(x, ...)
}
\arguments{
\item{x}{A ROC coordinates object.
A simple data.frame containing respective `x` and `y` coordinates.
The coordinates should be ROC space compliant.}

\item{...}{Other parameters passed to base plot function.}
}
\value{
plot of a ROC curve for given `roc_curve`
}
\description{
Plot ROC points using provided coordinates.
}
