% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_predicted_values.R
\name{positive_predicted_values}
\alias{positive_predicted_values}
\title{Positive Predicted Values}
\usage{
positive_predicted_values(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Positive Predicted Values
}
\description{
Compute Positive Predicted Values using confusion matrix.
Indicates what is the proportion of examples that truly belong to class i, from among all the examples assigned to (or classified as) class i.
}
