% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_coordinates.R
\name{pr_coordinates}
\alias{pr_coordinates}
\title{Precision–Recall (PR)}
\usage{
pr_coordinates(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
matrix of PR coordinates for given confusion matrix
}
\description{
Get Precision–Recall (PR) coordinates.
Precision–recall Curves, sometimes abbreviated as PR curves, are similar to
ROC curves and lift charts in that they explore the trade-off between the well-classified
positive examples and the number of misclassified negative examples.
}
