% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_curve.R
\name{pr_curve}
\alias{pr_curve}
\title{Precision–Recall (PR)}
\usage{
pr_curve(y_real, y_predicted, threshold_range = c(0, 1), threshold_step = 0.01)
}
\arguments{
\item{y_real}{Observed values to compare with in matrix format.}

\item{y_predicted}{Predicte values (probabiblities by class).}

\item{threshold_range}{Decision boundary threshold range.
Defaults to `c(0, 1)` for basic probabilistic classifier, but other values may be used.}

\item{threshold_step}{Step to use over the boundary threshold range.
Defaults to 0.01 for binary classification.}
}
\value{
matrix of PR curve coordinates for given probabilistic classifier output.
}
\description{
Get Precision–Recall (PR) curve construction.
}
