% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer value of Recall
}
\description{
Compute Recall using confusion matrix.
"Recall" is another term to determine "True Positive Ratio" (TPR).
In this package the `recall()` function is defined as a wrapper for `true_positive_ratio()`.
}
