% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root-Mean-Square Error (RMSE)}
\usage{
rmse(y_real, y_predicted)
}
\arguments{
\item{y_real}{Observed values (integers) to compare with
(in matrix format for multiclass classification).}

\item{y_predicted}{Predicte values (probabiblities by class).}
}
\value{
integer value of Root-Mean-Square Error (RMSE)
}
\description{
Compute Root-Mean-Square Error (RMSE) using confusion matrix.
The squared loss, in a sense, quantifies the error (or alternatively closeness) of the predicted label to the true label.
When specialized to the case of probabilistic classifiers, this then can be interpreted as a reliability measure.
}
