% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_coordinates.R
\name{roc_coordinates}
\alias{roc_coordinates}
\title{Receiver Operating Characteristic (ROC)}
\usage{
roc_coordinates(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
matrix of ROC coordinates for given confusion matrix
}
\description{
Get Receiver Operating Characteristic (ROC) coordinates.
An ROC curve is a plot in which the horizontal axis (the x axis) denotes the false-positive rate.
The vertical axis (the y axis) denotes the true-positive rate of a classifier.
ROC analysis studies the relationship between the sensitivity and the specificity of the classifier.
It makes sence in the context of binary classification.
}
