% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve.R
\name{roc_curve}
\alias{roc_curve}
\title{Receiver Operating Characteristic (ROC)}
\usage{
roc_curve(
  y_real,
  y_predicted,
  threshold_range = c(0, 1),
  threshold_step = 0.01
)
}
\arguments{
\item{y_real}{Observed values to compare with in matrix format.}

\item{y_predicted}{Predicte values (probabiblities by class).}

\item{threshold_range}{Decision boundary threshold range.
Defaults to `c(0, 1)` for basic probabilistic classifier, but other values may be used.}

\item{threshold_step}{Step to use over the boundary threshold range.
Defaults to 0.01 for binary classification.}
}
\value{
matrix of ROC curve coordinates for given probabilistic classifier output.
}
\description{
Get Receiver Operating Characteristic (ROC) curve construction.
An ROC curve is a plot in which the horizontal axis (the x axis) denotes the false-positive rate.
The vertical axis (the y axis) denotes the true-positive rate of a classifier.
ROC analysis studies the relationship between the sensitivity and the specificity of the classifier.
It makes sence in the context of binary classification.
}
