% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_coefficient.R
\name{s_coefficient}
\alias{s_coefficient}
\title{S Coefficient}
\usage{
s_coefficient(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer S Coefficient value.
}
\description{
Calculate S Coefficient using confusion matrix.
For more information see Bennett et al. (1954).
}
