% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scott_pi.R
\name{scott_pi}
\alias{scott_pi}
\title{Scott's Pi}
\usage{
scott_pi(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
integer Scott's Pi value.
}
\description{
Calculate Scott's Pi using confusion matrix.
It estimates the chance that a label (positive or negative) is assigned given a random instance irrespective of the label-assigning process.
For more information see Scott (1955)
}
