% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of Sensitivity values.
}
\description{
Calculate Sensitivity using confusion matrix.
"Sensitivity" is just another term to designate "True Positive Ratio".
The term has its origin in the medical domain, in which the metric is typically used to study the effectiveness of a clinical test in detecting a disease (Japkowitcz, 2011).
In the multiclass scenario, sensitivity measurements essentially study the accuracy of the classifier over individual classes.
In this package the function is just a wrapper for `true_negative_ratio()`.
}
