% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of Specificity values.
}
\description{
Calculate Specificity using confusion matrix.
"Specificity" is just another term to designate "True Negative Ratio".
This is the complement metric to "Sensitivity" metric.
In this package the function is just a wrapper for `true_negative_ratio()`.
}
