% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_negative.R
\name{true_negative}
\alias{true_negative}
\title{True Negative}
\usage{
true_negative(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of True Negatives count
}
\description{
Count True Negatives using confusion matrix.
}
