% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_negative_rate.R
\name{true_negative_rate}
\alias{true_negative_rate}
\title{True Negative Rate}
\usage{
true_negative_rate(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of True Negatives Rate values.
}
\description{
Calculate True Negative Rate using confusion matrix.
}
