% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_positive.R
\name{true_positive}
\alias{true_positive}
\title{True Positive}
\usage{
true_positive(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of True Positives count
}
\description{
Count True Positives using confusion matrix.
}
