% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_positive_rate.R
\name{true_positive_rate}
\alias{true_positive_rate}
\title{True Positive Rate}
\usage{
true_positive_rate(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix to be used.}
}
\value{
vector of True Positives Rate values.
}
\description{
Calculate True Positive Rate using confusion matrix.
The proportion of the examples of some class i of interest actually assigned to class i by the learning algorithm.
}
