% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_coordinates.R
\name{plot.roc_coordinates}
\alias{plot.roc_coordinates}
\title{Plot ROC coordinates}
\usage{
\method{plot}{roc_coordinates}(roc_coordinates, ...)
}
\arguments{
\item{roc_coordinates}{A ROC coordinates object.
A simple data.frame containing respective `x` and `y` coordinates.
The coordinates should be ROC space compliant.}
}
\value{
plot of ROC coordinates for given coordinates
}
\description{
Plot ROC points using provided coordinates.
}
